[REFACTOR] 엔티티 네이밍 수정
