+++ src/main/java/com/sports/server/command/team/domain/LogoImageDeletedEvent.java
@@ -1,4 +1,4 @@
-package com.sports.server.command.leagueteam.domain;
+package com.sports.server.command.team.domain;
 
 public record LogoImageDeletedEvent(
         String logoImageUrl
