+++ src/main/java/com/sports/server/command/team/dto/LeagueTeamRequest.java
@@ -1,4 +1,4 @@
-package com.sports.server.command.leagueteam.dto;
+package com.sports.server.command.team.dto;
 
 import com.sports.server.command.league.domain.League;
 import com.sports.server.command.leagueteam.domain.LeagueTeam;
