+++ src/test/java/com/sports/server/support/DocumentationTest.java
@@ -13,8 +13,8 @@ import com.sports.server.command.game.application.LineupPlayerService;
 import com.sports.server.command.game.presentation.GameController;
 import com.sports.server.command.league.application.LeagueService;
 import com.sports.server.command.league.presentation.LeagueController;
-import com.sports.server.command.leagueteam.application.LeagueTeamService;
-import com.sports.server.command.leagueteam.presentation.LeagueTeamController;
+import com.sports.server.command.team.application.LeagueTeamService;
+import com.sports.server.command.team.presentation.LeagueTeamController;
 import com.sports.server.command.member.domain.Member;
 import com.sports.server.command.member.domain.MemberRepository;
 import com.sports.server.command.report.application.ReportService;
